Input & output plugins for kafka
