+++ package.json
@@ -34,13 +34,18 @@
   "dependencies": {
     "alasql": "^0.3.3",
     "bunyan": "^1.8.5",
+    "codecov": "^2.3.0",
+    "coveralls": "^2.13.1",
     "docopt": "^0.6.2",
     "elasticsearch": "^13.0.1",
     "eventemitter3": "^1.2.0",
+    "istanbul-diff": "^1.1.4",
     "js-yaml": "^3.3.1",
-    "kafka-node": "^2.2.2",
+    "kafka-node": "^2.2.3",
     "maxmind-db-reader": "^0.2.1",
     "moment": "^2.10.3",
+    "node-rdkafka": "^2.0.0",
+    "nyc": "^11.2.1",
     "object-sizeof": "^1.1.1",
     "simple-get": "^2.4.0",
     "split2": "^2.1.0",
